Actualizacion de boleta y pago

+++ app/controllers/boleta/D_boleta.php
@@ -4,8 +4,17 @@ include('../../../config/conexion.php');
 $codigo = $_POST['txt_id'];
 $volver = $_POST['txt_volver'];
 
-$sql ="DELETE FROM boleta WHERE id_bo = $codigo";
-mysqli_query($cn,$sql);
+
+
+$sqlp ="DELETE FROM pago  WHERE id_bo = $codigo";
+mysqli_query($cn,$sqlp);
+if ($sqlp) {
+
+    $sql ="DELETE FROM boleta WHERE id_bo = $codigo";
+    mysqli_query($cn,$sql);
+    echo '<script>window.location.href = "../../../boleta.php?id=' . $volver . '";</script>';
+}
+
 
 echo '<script>window.location.href = "../../../boleta.php?id=' . $volver . '";</script>';
 ?>
